Optimizations in formatting
- move range and tooltip files into single js file.
- making formatting lazy since the first time API is called it takes around 30-50 ms
